services/backend/requirements.txt
# LangGraph & LangChain
langgraph==0.2.53
langchain==0.3.13
langchain-anthropic==0.3.0
langchain-openai==0.2.14
langchain-google-genai==2.0.8

# Chainlit (UI) - 최신 버전으로 업그레이드
chainlit>=2.0.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10

# MCP (Model Context Protocol)
mcp==1.1.2
httpx==0.28.1
kubernetes==31.0.0
psycopg2-binary==2.9.10
GitPython==3.1.43

# YAML
PyYAML==6.0.2

# Utilities
python-dotenv==1.0.1
redis==5.2.1
aioredis==2.0.1
httpx==0.28.1

